Hashin, bloomfilter, count-min sketch
A new class now contains the necessary methods to generate a number of
hash functions and a helper to return a number of hash values for each
of the functions and a given key.

This class is used in hash-based datastructures like Count-Min Sketches
or Bloom Filters.

A bloom filter is added on top of the hashing method. The Count-Min
sketch has been adopted to the new hashing method.